Use an epsilon when comparing floating point.
This uses an epsilon when comparing floating point numbers. Many costs
are almost identical, and the solver can spend a long time optimising
almost equal decisions. This dramatically decreases the worst-case
run-time (15s vs 4s in one test case) for a very small change in the
solution.

Notably, on my test case, the solution is different by one route swap
but the cost is the same to 11 decimal places.